my_app/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
tensorflow==2.17.0
numpy==1.21.0
Pillow==9.0.1